Some reformats and cleanup warnings; Mostly allow posted update portion to update only that portion on paint callback

+++ src/vidlib/puregl/vidlib.c
@@ -621,7 +621,7 @@ LRESULT CALLBACK
 					{
 						if( l.flags.bLogKeyEvent )
 							lprintf( WIDE("already dispatched, delay it.") );
-						EnqueLink( &hVideo->pInput, (POINTER)key );
+						EnqueLink( &hVideo->pInput, (POINTER)(uintptr_t)key );
 					}
 					else
 					{
@@ -678,7 +678,7 @@ LRESULT CALLBACK
 									lprintf( WIDE( "lost active window." ) );
 								break;
 							}
-							key = (uint32_t)DequeLink( &hVideo->pInput );
+							key = (uint32_t)(uintptr_t)DequeLink( &hVideo->pInput );
 							if( l.flags.bLogKeyEvent )
 								lprintf( WIDE( "key from deque : %p" ), key );
 						} while( key );
@@ -915,7 +915,7 @@ LRESULT CALLBACK
 						hVideo->flags.event_dispatched = 1;
 						//lprintf( WIDE("Dispatched KEY!") );
 						if( hVideo->flags.key_dispatched )
-							EnqueLink( &hVideo->pInput, (POINTER)key );
+							EnqueLink( &hVideo->pInput, (POINTER)(uintptr_t)key );
 						else
 						{
 							hVideo->flags.key_dispatched = 1;
@@ -958,7 +958,7 @@ LRESULT CALLBACK
 								}
 								if( FindLink( &l.pActiveList, hVideo ) == INVALID_INDEX )
 									break;
-								key = (uint32_t)DequeLink( &hVideo->pInput );
+								key = (uint32_t)(uintptr_t)DequeLink( &hVideo->pInput );
 							} while( key );
 							hVideo->flags.key_dispatched = 0;
 						}
